// src/enums/nameSuffix.ts
export enum NameSuffix {
    Junior = "Junior",
    Senior = "Senior",
    First = "First",
    Second = "Second",
    Third = "Third",
    Fourth = "Fourth",
    Fifth = "Fifth",
    Sixth = "Sixth",
    Seventh = "Seventh",
    Eighth = "Eighth",
    Ninth = "Ninth",
    Unknown = "Unknown"
  }
  